# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Nobellum Score Sheet API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and PIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current session info
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Session info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /scores/submit:
    post:
      operationId: submitScores
      tags: [scores]
      summary: Submit score rows (upsert by dedupe key)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitScoresRequest"
      responses:
        "200":
          description: Scores submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmitScoresResponse"
        "403":
          description: Access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /scores:
    get:
      operationId: getScores
      tags: [scores]
      summary: Get scores with filters
      security:
        - bearerAuth: []
      parameters:
        - name: programId
          in: query
          required: true
          schema:
            type: string
        - name: cohortId
          in: query
          required: true
          schema:
            type: string
        - name: round
          in: query
          required: true
          schema:
            type: string
        - name: startupId
          in: query
          schema:
            type: integer
        - name: judgeId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Scores list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Score"
  /committee/analytics:
    get:
      operationId: getCommitteeAnalytics
      tags: [committee]
      summary: Get committee analytics (rankings, judge analytics, matrix)
      security:
        - bearerAuth: []
      parameters:
        - name: programId
          in: query
          required: true
          schema:
            type: string
        - name: cohortId
          in: query
          required: true
          schema:
            type: string
        - name: round
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Analytics data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsResponse"
        "403":
          description: Access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /committee/advance-preview:
    get:
      operationId: getAdvancePreview
      tags: [committee]
      summary: Preview auto-advance decisions
      security:
        - bearerAuth: []
      parameters:
        - name: programId
          in: query
          required: true
          schema:
            type: string
        - name: cohortId
          in: query
          required: true
          schema:
            type: string
        - name: fromRound
          in: query
          required: true
          schema:
            type: string
        - name: toRound
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Advance preview
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdvancePreviewItem"
        "403":
          description: Access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        pin:
          type: string
      required:
        - email
        - pin
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        role:
          type: string
        user:
          $ref: "#/components/schemas/UserInfo"
        allowedAssignments:
          type: array
          items:
            $ref: "#/components/schemas/AllowedAssignment"
      required:
        - token
        - role
        - user
        - allowedAssignments
    UserInfo:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
        judgeId:
          type: integer
          nullable: true
      required:
        - id
        - email
        - name
        - role
    AllowedAssignment:
      type: object
      properties:
        programId:
          type: string
        cohortId:
          type: string
      required:
        - programId
        - cohortId
    MeResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserInfo"
        allowedAssignments:
          type: array
          items:
            $ref: "#/components/schemas/AllowedAssignment"
      required:
        - user
        - allowedAssignments
...
    Judge:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        title:
          type: string
          nullable: true
        active:
          type: boolean
      required:
        - id
        - name
        - email
        - active
...
    Score:
      type: object
      properties:
        id:
          type: integer
        programId:
          type: string
        cohortId:
          type: string
        roundName:
          type: string
        startupId:
          type: integer
        judgeId:
          type: integer
        criterionId:
          type: string
        score:
          type: integer
        weight:
          type: number
        comment:
          type: string
          nullable: true
        source:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - programId
        - cohortId
        - roundName
        - startupId
        - judgeId
        - criterionId
        - score
        - weight
        - source
        - createdAt
        - updatedAt
    ScoreRow:
      type: object
      properties:
        programId:
          type: string
        cohortId:
          type: string
        roundName:
          type: string
        startupId:
          type: integer
        judgeId:
          type: integer
        criterionId:
          type: string
        score:
          type: integer
        weight:
          type: number
        comment:
          type: string
          nullable: true
      required:
        - programId
        - cohortId
        - roundName
        - startupId
        - judgeId
        - criterionId
        - score
        - weight
    SubmitScoresRequest:
      type: object
      properties:
        rows:
          type: array
          items:
            $ref: "#/components/schemas/ScoreRow"
      required:
        - rows
    SubmitScoresResponse:
      type: object
      properties:
        submitted:
          type: integer
      required:
        - submitted
    StartupRanking:
      type: object
      properties:
        startupId:
          type: integer
        startupName:
          type: string
        avgPct:
          type: number
        sdPct:
          type: number
        judgeCount:
          type: integer
      required:
        - startupId
        - startupName
        - avgPct
        - sdPct
        - judgeCount
    JudgeAnalytic:
      type: object
      properties:
        judgeId:
          type: integer
        judgeName:
          type: string
        meanPct:
          type: number
        sdPct:
          type: number
        bias:
          type: number
        startupsScored:
          type: integer
      required:
        - judgeId
        - judgeName
        - meanPct
        - sdPct
        - bias
        - startupsScored
    MatrixCell:
      type: object
      properties:
        startupId:
          type: integer
        startupName:
          type: string
        judgeScores:
          type: object
          additionalProperties:
            type: number
      required:
        - startupId
        - startupName
        - judgeScores
    AnalyticsResponse:
      type: object
      properties:
        rankings:
          type: array
          items:
            $ref: "#/components/schemas/StartupRanking"
        judgeAnalytics:
          type: array
          items:
            $ref: "#/components/schemas/JudgeAnalytic"
        matrix:
          type: array
          items:
            $ref: "#/components/schemas/MatrixCell"
        judges:
          type: array
          items:
            $ref: "#/components/schemas/Judge"
      required:
        - rankings
        - judgeAnalytics
        - matrix
        - judges
    AdvancePreviewItem:
      type: object
      properties:
        startupId:
          type: integer
        startupName:
          type: string
        avgPct:
          type: number
        judgeCount:
          type: integer
        decision:
          type: string
          enum: [advance, hold]
        thresholdPct:
          type: number
        minJudges:
          type: integer
      required:
        - startupId
        - startupName
        - avgPct
        - judgeCount
        - decision
        - thresholdPct
        - minJudges
...